Collapse the dead client-generator base class into ClientsGenerator
BaseClientsGenerator had exactly one subclass, and the only method that
justified the split - its generate_function - rendered templates
(get_method.jinja2, post_method.jinja2) that do not exist anywhere in
the repository, so it could never run. The base/subclass split also
forced the FrameworkHTTPPlaceholder shim, whose only job was to carry a
dict between methods of what is logically one class.

Everything now lives in a single ClientsGenerator, with the status-code
bundle as a plain attribute. Also removed along the way, all verified
unused by code or templates:

- ParametersResponse.get_path_args_as_string / get_required_path_args_as_string / has_query_args
- shared/utils create_query_args and create_query_args_with_mapping
  (only callers were the dead generate_function)
- the new_unions template variable (no template references it)

Regenerating the test clients produces byte-identical output.
